docs: renumber Scrapling guide to 07 and switch to a single-file example
